requirements.txt
django_recaptcha==3.0.0
matplotlib==3.8.0
Pillow==10.1.0
Pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
captcha==0.5.0
